Split HBase spark/non spark code
