+++ build.sbt
@@ -107,7 +107,8 @@ lazy val root = Project("geotrellis", file(".")).
     geomesa,
     geotools,
     geowave,
-    hbase,
+    `hbase-store`,
+    `hbase-spark`,
     macros,
     proj4,
     raster,
@@ -228,7 +229,7 @@ lazy val cassandra = project
   .settings(commonSettings)
   .settings(Settings.cassandra)
 
-lazy val hbase = project
+lazy val `hbase-store` = project
   .dependsOn(
     spark % "compile->compile;test->test", // <-- spark-testkit update should simplify this
     `spark-testkit` % Test
@@ -237,9 +238,19 @@ lazy val hbase = project
   .settings(Settings.hbase)
   .settings(projectDependencies := { Seq((projectID in spark).value.exclude("com.google.protobuf", "protobuf-java")) })
 
-lazy val `spark-etl` = Project(id = "spark-etl", base = file("spark-etl")).
-  dependsOn(spark, s3, accumulo, cassandra, hbase).
-  settings(commonSettings)
+lazy val `hbase-spark` = project
+  .dependsOn(
+    `hbase-store` % "compile->compile;test->test",
+    spark % "compile->compile;test->test", // <-- spark-testkit update should simplify this
+    `spark-testkit` % Test
+  )
+  .settings(commonSettings) // HBase depends on its own protobuf version
+  .settings(Settings.hbase)
+  .settings(projectDependencies := { Seq((projectID in spark).value.exclude("com.google.protobuf", "protobuf-java")) })
+
+lazy val `spark-etl` = Project(id = "spark-etl", base = file("spark-etl"))
+  .dependsOn(spark, s3, accumulo, cassandra, `hbase-store`, `hbase-spark`)
+  .settings(commonSettings)
   .settings(Settings.`spark-etl`)
 
 lazy val `spark-pipeline` = Project(id = "spark-pipeline", base = file("spark-pipeline")).
@@ -279,7 +290,7 @@ lazy val util = project
   .settings(Settings.util)
 
 lazy val `doc-examples` = project
-  .dependsOn(spark, s3, accumulo, cassandra, hbase, spark, `spark-testkit`, `spark-pipeline`)
+  .dependsOn(spark, s3, accumulo, cassandra, `hbase-store`, `hbase-spark`, spark, `spark-testkit`, `spark-pipeline`)
   .settings(commonSettings)
   .settings(Settings.`doc-examples`)
