+++ hbase-store/src/main/resources/META-INF/services/geotrellis.spark.io.AttributeStoreProvider
@@ -1,0 +1,1 @@
+geotrellis.spark.io.hbase.HBaseLayerProvider
