+++ hbase-store/src/main/resources/META-INF/services/geotrellis.spark.io.CollectionLayerReaderProvider
@@ -1,0 +1,1 @@
+geotrellis.spark.io.hbase.HBaseLayerProvider
